backend/requirements.txt
PyMySQL==0.6.6
peewee==2.5.1
tornado==4.1
isodate==0.5.1
requests==2.7.0
soundcloud==0.4.1
spotipy==2.3.7
dogpile.cache==0.5.7

google-api-python-client==1.4.0
